modules/vibration_sensor/requirements.txt
RPi.GPIO==0.7.0

